Move suicide logic into its own file
